+++ packages/server/src/internals/graphql/utils.ts
@@ -121,24 +121,12 @@ export function assertStringIsNotEmpty(args: { [key: string]: unknown }) {
   }
 }
 
-export function checkRoleInstanceCompatibility(
-  role: Role,
-  instance: Instance | null
-) {
-  if (
-    [Role.Guest, Role.Login, Role.Sysadmin].includes(role) &&
-    isInstance(instance)
-  ) {
-    throw new UserInputError('This role cannot be scoped.')
-  }
-  if (
-    ![Role.Guest, Role.Login, Role.Sysadmin].includes(role) &&
-    !isInstance(instance)
-  ) {
-    throw new UserInputError("This role can't have a global scope.")
-  }
+export function isGlobalRole(role: Role): boolean {
+  return [Role.Guest, Role.Login, Role.Sysadmin].includes(role)
 }
 
 export function generateRole(role: Role, instance: Instance | null) {
-  return isInstance(instance) ? `${instance}_${role}` : role
+  if (isGlobalRole(role)) return role
+  if (isInstance(instance)) return `${instance}_${role}`
+  else throw new UserInputError('This role needs an instance')
 }